+++ src/Makefile.am
@@ -82,6 +82,7 @@ forked_daapd_SOURCES = main.c \
 	filescanner.c filescanner.h \
 	filescanner_ffmpeg.c filescanner_playlist.c \
 	filescanner_smartpl.c $(ITUNES_SRC) \
+	library.c library.h \
 	mdns_avahi.c mdns.h \
 	remote_pairing.c remote_pairing.h \
 	avio_evbuffer.c avio_evbuffer.h \
